New page displaying Blitz Results
I figured that the display of the words submitted, the UI allowing for the words to be deleted and the tabulation of scores is gonna be pretty big esp if we are gonna do real analysis like clusters missed etc. In addition to show all the words and allow users to delete them scrollview has to be in place, and scroll view affects a lot of things  so standalone is best.